// frontend/src/interfaces/Bookmark.ts
export interface BookmarkItem {
  id: string;
  name: string;
  imageUrl: string;
  description: string;
  rating: number;
}

// export interface bookmarkItem {
//   bookMarkId: number;
//   categoryId: number;
//   courseId: number;
//   courseTitle: string;
//   destination: string;
//   content: string;
//   isPublished: string;
//   courseImgUrl: string;
//   createdDate: string;
//   modifiedDate: string;
// }
// export interface bookmarkData {
//   result: string;
//   userId: string;
//   data: bookmarkItem[];
// }